Fix the nil cast on error for podlist and refactor pod retry call (#266)
* Fix the nil cast on error for podlist and refactor pod retry call

In pod retry the object returned from listing was used to cast even when error was returned. This PR
fixes this issue. The retry call has been refactored to account for other error conditions and simplify
the usage. This PR introduces getPodList call and then a retry call built on top of that.

* Address few review comments

* Address review coments

* Address review comments